refactor(core): merge SAML assertion handler API to authn API
