+++ packages/core/src/routes/authn.test.ts
@@ -1,9 +1,13 @@
+import { ConnectorType } from '@logto/connector-kit';
 import { pickDefault, createMockUtils } from '@logto/shared/esm';
 
 import { mockRole } from '#src/__mocks__/index.js';
 import RequestError from '#src/errors/RequestError/index.js';
 import Libraries from '#src/tenants/Libraries.js';
+import { createMockProvider } from '#src/test-utils/oidc-provider.js';
 import { MockTenant } from '#src/test-utils/tenant.js';
+
+import { mockConnector, mockMetadata, mockLogtoConnector } from '../__mocks__/connector.js';
 
 const { jest } = import.meta;
 const { mockEsmWithActual } = createMockUtils(jest);
@@ -14,12 +18,52 @@ const { verifyBearerTokenFromRequest } = await mockEsmWithActual(
     verifyBearerTokenFromRequest: jest.fn(),
   })
 );
+const validateSamlAssertion = jest.fn();
+
+const mockSamlLogtoConnector = {
+  dbEntry: { ...mockConnector, connectorId: 'saml', id: 'saml_connector' },
+  metadata: { ...mockMetadata, isStandard: true, id: 'saml', target: 'saml' },
+  type: ConnectorType.Social,
+  ...mockLogtoConnector,
+  validateSamlAssertion,
+};
+
+const socialsLibraries = {
+  getConnector: jest.fn(async (connectorId: string) => {
+    if (connectorId !== 'saml_connector') {
+      throw new RequestError({
+        code: 'entity.not_found',
+        connectorId,
+        status: 404,
+      });
+    }
+
+    return mockSamlLogtoConnector;
+  }),
+};
+
+const baseProviderMock = {
+  params: {},
+  jti: 'jti',
+  client_id: 'client_id',
+};
+
+// Const samlAssertionHandlerRoutes = await pickDefault(import('./authn/saml.js'));
+// const tenantContext = new MockTenant(
+//   createMockProvider(jest.fn().mockResolvedValue(baseProviderMock)),
+//   undefined,
+//   { socials: socialsLibraries }
+// );
 
 const usersLibraries = {
   findUserRoles: jest.fn(async () => [mockRole]),
 } satisfies Partial<Libraries['users']>;
 
-const tenantContext = new MockTenant(undefined, {}, { users: usersLibraries });
+const tenantContext = new MockTenant(
+  createMockProvider(jest.fn().mockResolvedValue(baseProviderMock)),
+  undefined,
+  { users: usersLibraries, socials: socialsLibraries }
+);
 const { createRequester } = await import('#src/utils/test-utils.js');
 const request = createRequester({
   anonymousRoutes: await pickDefault(import('#src/routes/authn.js')),
@@ -121,5 +165,35 @@ describe('authn route for Hasura', () => {
         [keys.hasuraRole]: mockUnauthorizedRole,
       });
     });
+  });
+});
+
+describe('authn route for SAML', () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('POST /authn/saml/non_saml_connector should throw 404', async () => {
+    const response = await request.post('/authn/saml/non_saml_connector');
+    expect(response.status).toEqual(404);
+  });
+
+  it('POST /authn/saml/saml_connector should throw when `RelayState` missing', async () => {
+    const response = await request.post('/authn/saml/saml_connector').send({
+      SAMLResponse: 'saml_response',
+    });
+    expect(response.status).toEqual(500);
+  });
+
+  it('POST /authn/saml/saml_connector', async () => {
+    await request.post('/authn/saml/saml_connector').send({
+      SAMLResponse: 'saml_response',
+      RelayState: 'relay_state',
+    });
+    expect(validateSamlAssertion).toHaveBeenCalledWith(
+      { body: { RelayState: 'relay_state', SAMLResponse: 'saml_response' } },
+      expect.anything(),
+      expect.anything()
+    );
   });
 });
